move fc to phi and try to fix bug
move fc to phi

try to fix bug
