+++ paddle/fluid/pir/dialect/op_generator/ops_api_gen.py
@@ -86,6 +86,7 @@ NEED_GEN_STATIC_ONLY_APIS = [
     'fusion_seqexpand_concat_fc',
     'fused_attention',
     'fused_feedforward',
+    'fc',
 ]
 
 NO_NEED_GEN_STATIC_ONLY_APIS = [
