+++ test/legacy_test/test_fc_op.py
@@ -84,7 +84,7 @@ class TestFCOp(OpTest):
         }
 
     def test_check_output(self):
-        self.check_output()
+        self.check_output(check_dygraph=False)
 
 
 class TestFCOpNoBias1(TestFCOp):
